+++ dimas/src/prelude.rs
@@ -5,11 +5,11 @@ pub use crate::agent::Agent;
 pub use crate::config::Config;
 pub use crate::context::Context;
 pub use crate::error::{Error, Result};
-pub use crate::message::Message;
+//pub use crate::message::Message;
 // endregeion:  --- modules
 
 // region:    --- types
-// Generic wrapper tuple struct for newtype pattern
+/// Generic wrapper tuple struct for newtype pattern
 #[repr(transparent)]
 pub struct Wrap<T>(pub T);
 // endregion: --- types